Update Readme.md
Justin and I both worked on Constant. I think he and maybe someone else also contributed to Integer? Log is taking quite a bit of work, so I thought I might go ahead and add that just in case. It probably doesn't matter, but I think we should try to squeeze every point out of this damn thing as possible.