Adding WolframRequest
Pulled it out and made a new class for the request. Comments inbound.
The files aren’t included anywhere, so everyone should still be able to
compile. If you need to use it, expect a string back with the result in
Wolfram Alpha’s notation. If you use the request, you will need to
handle the parsing.